+++ src/Omnipay/AuthorizeNet/SIMGateway.php
@@ -24,6 +24,25 @@ class SIMGateway extends AIMGateway
         return 'Authorize.Net SIM';
     }
 
+    public function getDefaultParameters()
+    {
+        $parameters = parent::getDefaultParameters();
+        $parameters['hashSecret'] = '';
+
+        return $parameters;
+    }
+
+    public function getHashSecret()
+    {
+        return $this->getParameter('hashSecret');
+    }
+
+    public function setHashSecret($value)
+    {
+        return $this->setParameter('hashSecret', $value);
+    }
+
+
     public function authorize(array $parameters = array())
     {
         return $this->createRequest('\Omnipay\AuthorizeNet\Message\SIMAuthorizeRequest', $parameters);