+++ src/Omnipay/MultiSafepay/Message/CompletePurchaseResponse.php
@@ -50,13 +50,13 @@ class CompletePurchaseResponse extends AbstractResponse
     }
 
     /**
-     * Is the payment cancelled?
+     * Is the payment canceled?
      *
      * @return boolean
      */
-    public function isCancelled()
+    public function isCanceled()
     {
-        return isset($this->data->ewallet->status) && 'void' === (string) $this->data->ewallet->status;
+        return isset($this->data->ewallet->status) && 'canceled' === (string) $this->data->ewallet->status;
     }
 
     /**