+++ src/Omnipay/Stripe/Message/FetchTransactionRequest.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of the Omnipay package.
+ *
+ * (c) Adrian Macneil <adrian@adrianmacneil.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Omnipay\Stripe\Message;
+
+/**
+ * Stripe Fetch Transaction Request
+ */
+class FetchTransactionRequest extends AbstractRequest
+{
+    public function getData()
+    {
+        $this->validate('transactionReference');
+
+        $data = array();
+
+        return $data;
+    }
+
+    public function getEndpoint()
+    {
+        return $this->endpoint.'/charges/'.$this->getTransactionReference();
+    }
+}