+++ src/Omnipay/TargetPay/IdealGateway.php
@@ -1,0 +1,63 @@
+<?php
+
+/*
+ * This file is part of the Omnipay package.
+ *
+ * (c) Adrian Macneil <adrian@adrianmacneil.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Omnipay\TargetPay;
+
+use Omnipay\TargetPay\Message\CompletePurchaseRequest;
+use Omnipay\TargetPay\Message\FetchIssuersRequest;
+
+/**
+ * TargetPay iDEAL gateway.
+ *
+ * @link https://www.targetpay.com/docs/TargetPay_iDEAL_V1.0_nl.pdf
+ */
+class IdealGateway extends AbstractGateway
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'TargetPay iDEAL';
+    }
+
+    /**
+     * Retrieve iDEAL issuers.
+     *
+     * @param array $parameters An array of options
+     *
+     * @return FetchIssuersRequest
+     */
+    public function fetchIssuers(array $parameters = array())
+    {
+        return $this->createRequest('\Omnipay\TargetPay\Message\FetchIssuersRequest', $parameters);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function purchase(array $parameters = array())
+    {
+        return $this->createRequest('\Omnipay\TargetPay\Message\IdealPurchaseRequest', $parameters);
+    }
+
+    /**
+     * Complete a purchase.
+     *
+     * @param array $parameters An array of options
+     *
+     * @return CompletePurchaseRequest
+     */
+    public function completePurchase(array $parameters = array())
+    {
+        return $this->createRequest('\Omnipay\TargetPay\Message\IdealCompletePurchaseRequest', $parameters);
+    }
+}