+++ tests/Omnipay/CardSave/GatewayTest.php
@@ -32,6 +32,9 @@ class GatewayTest extends GatewayTestCase
                 'expiryMonth' => '12',
                 'expiryYear' => '2016',
                 'cvv' => '123',
+                'issueNumber' => '5',
+                'startMonth' => '4',
+                'startYear' => '2013',
             )),
         );
     }