+++ tests/Omnipay/PayPal/Message/ExpressAuthorizeResponseTest.php
@@ -18,7 +18,8 @@ class ExpressAuthorizeResponseTest extends TestCase
     public function testConstruct()
     {
         // response should decode URL format data
-        $response = new ExpressAuthorizeResponse('example=value&foo=bar');
+        $response = new ExpressAuthorizeResponse($this->getMockRequest(), 'example=value&foo=bar');
+
         $this->assertEquals(array('example' => 'value', 'foo' => 'bar'), $response->getData());
     }
 
@@ -27,10 +28,11 @@ class ExpressAuthorizeResponseTest extends TestCase
         $httpResponse = $this->getMockHttpResponse('ExpressPurchaseSuccess.txt');
         $response = new ExpressAuthorizeResponse($this->getMockRequest(), $httpResponse->getBody());
 
-        $this->assertTrue($response->isSuccessful());
-        $this->assertSame('EC-42721413K79637829', $response->getExpressRedirectToken());
-        $this->assertNull($response->getTransactionReference());
+        $this->assertFalse($response->isSuccessful());
+        $this->assertSame('EC-42721413K79637829', $response->getTransactionReference());
         $this->assertNull($response->getMessage());
+        $this->assertNull($response->getRedirectData());
+        $this->assertSame('GET', $response->getRedirectMethod());
     }
 
     public function testExpressPurchaseFailure()
@@ -39,7 +41,7 @@ class ExpressAuthorizeResponseTest extends TestCase
         $response = new ExpressAuthorizeResponse($this->getMockRequest(), $httpResponse->getBody());
 
         $this->assertFalse($response->isSuccessful());
-        $this->assertNull($response->getExpressRedirectToken());
+        $this->assertNull($response->getTransactionReference());
         $this->assertNull($response->getTransactionReference());
         $this->assertSame('This transaction cannot be processed. The amount to be charged is zero.', $response->getMessage());
     }