+++ tests/Omnipay/SagePay/Message/DirectAuthorizeRequestTest.php
@@ -1,0 +1,55 @@
+<?php
+
+/*
+ * This file is part of the Omnipay package.
+ *
+ * (c) Adrian Macneil <adrian@adrianmacneil.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Omnipay\SagePay\Message;
+
+use Omnipay\TestCase;
+
+class DirectAuthorizeRequestTest extends TestCase
+{
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->request = new DirectAuthorizeRequest($this->getHttpClient(), $this->getHttpRequest());
+        $this->request->initialize(
+            array(
+                'amount' => '12.00',
+                'transactionId' => '123',
+                'card' => $this->getValidCard(),
+            )
+        );
+    }
+
+    public function testGetDataVisa()
+    {
+        $this->request->getCard()->setNumber('4929000000006');
+        $data = $this->request->getData();
+
+        $this->assertSame('visa', $data['CardType']);
+    }
+
+    public function testGetDataMastercard()
+    {
+        $this->request->getCard()->setNumber('5404000000000001');
+        $data = $this->request->getData();
+
+        $this->assertSame('mc', $data['CardType']);
+    }
+
+    public function testGetDataDinersClub()
+    {
+        $this->request->getCard()->setNumber('30569309025904');
+        $data = $this->request->getData();
+
+        $this->assertSame('dc', $data['CardType']);
+    }
+}