+++ tests/Omnipay/Stripe/Message/FetchTransactionTest.php
@@ -1,0 +1,52 @@
+<?php
+
+/*
+ * This file is part of the Omnipay package.
+ *
+ * (c) Adrian Macneil <adrian@adrianmacneil.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Omnipay\Stripe\Message;
+
+use Omnipay\TestCase;
+
+class FetchTransactionRequestTest extends TestCase
+{
+    public function setUp()
+    {
+        $this->request = new FetchTransactionRequest($this->getHttpClient(), $this->getHttpRequest());
+        $this->request->setTransactionReference('ch_29yrvk84GVDsq9');
+    }
+
+    public function testEndpoint()
+    {
+        $this->assertSame('https://api.stripe.com/v1/charges/ch_29yrvk84GVDsq9', $this->request->getEndpoint());
+    }
+
+    public function testSendSuccess()
+    {
+        $this->setMockHttpResponse('FetchTransactionSuccess.txt');
+        $response = $this->request->send();
+
+        $this->assertTrue($response->isSuccessful());
+        $this->assertFalse($response->isRedirect());
+        $this->assertSame('ch_29yrvk84GVDsq9', $response->getTransactionReference());
+        $this->assertNull($response->getCardReference());
+        $this->assertNull($response->getMessage());
+    }
+
+    public function testSendError()
+    {
+        $this->setMockHttpResponse('FetchTransactionFailure.txt');
+        $response = $this->request->send();
+
+        $this->assertFalse($response->isSuccessful());
+        $this->assertFalse($response->isRedirect());
+        $this->assertNull($response->getTransactionReference());
+        $this->assertNull($response->getCardReference());
+        $this->assertSame('No such charge: ch_29yrvk84GVDsq9fake', $response->getMessage());
+    }
+}