+++ tests/Omnipay/TargetPay/Message/FetchIssuersRequestTest.php
@@ -1,0 +1,52 @@
+<?php
+
+/*
+ * This file is part of the Omnipay package.
+ *
+ * (c) Adrian Macneil <adrian@adrianmacneil.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Omnipay\TargetPay\Message;
+
+use Omnipay\TestCase;
+
+class FetchIssuersRequestTest extends TestCase
+{
+    /**
+     * @var FetchIssuersRequest
+     */
+    private $request;
+
+    protected function setUp()
+    {
+        $this->request = new FetchIssuersRequest($this->getHttpClient(), $this->getHttpRequest());
+    }
+
+    /**
+     * @dataProvider issuersProvider
+     */
+    public function testSendSuccess($expected)
+    {
+        $this->setMockHttpResponse('FetchIssuersSuccess.txt');
+
+        $response = $this->request->send();
+
+        $this->assertTrue($response->isSuccessful());
+        $this->assertEquals($expected, $response->getIssuers());
+    }
+
+    public function issuersProvider()
+    {
+        return array(
+            array(
+                array(
+                    '0001' => 'Sample 1',
+                    '0002' => 'Sample 2',
+                ),
+            ),
+        );
+    }
+}